+++ runtime/dummychecking.js
@@ -1,0 +1,1 @@
+//TODO: Actually make functions here, lmao